+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/InvoiceLineAdjustments.js
@@ -1,0 +1,144 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import {
+  useIntl,
+  FormattedMessage,
+} from 'react-intl';
+import { Field } from 'redux-form';
+
+import {
+  Accordion,
+  Row,
+  Col,
+  Card,
+  IconButton,
+  RepeatableField,
+  TextField,
+  Checkbox,
+} from '@folio/stripes/components';
+import { TypeToggle } from '@folio/stripes-acq-components';
+
+import { AcceptedValuesField } from '../../../../../components';
+
+import {
+  onAdd,
+  onRemove,
+  getSubfieldName,
+} from '../../utils';
+import { TRANSLATION_ID_PREFIX } from '../../constants';
+import {
+  createOptionsList,
+  mappingProfileSubfieldShape,
+  okapiShape,
+  INOVOICE_ADJUSTMENTS_RELATION_TO_TOTAL_OPTIONS,
+} from '../../../../../utils';
+
+export const InvoiceLineAdjustments = ({
+  lineAdjustments,
+  initialFields,
+  setReferenceTables,
+  okapi,
+}) => {
+  const { formatMessage } = useIntl();
+
+  const relationToTotalList = createOptionsList(INOVOICE_ADJUSTMENTS_RELATION_TO_TOTAL_OPTIONS, formatMessage);
+
+  const renderLineAdjustment = (field, index) => {
+    const trashButton = (
+      <IconButton
+        data-test-repeatable-field-remove-item-button
+        icon="trash"
+        onClick={() => onRemove(index, lineAdjustments, 41, setReferenceTables, 'order')}
+        size="medium"
+        ariaLabel={formatMessage({ id: 'stripes-components.deleteThisItem' })}
+      />
+    );
+
+    const headerTitle = (
+      <FormattedMessage
+        id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.adjustments.sectionTitle`}
+        values={{ index: index + 1 }}
+      />
+    );
+
+    return (
+      <Card
+        headerEnd={trashButton}
+        headerStart={headerTitle}
+      >
+        <Row left="xs">
+          <Col xs={3}>
+            <Field
+              component={TextField}
+              label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.field.description`} />}
+              name={getSubfieldName(41, 0, index)}
+            />
+          </Col>
+          <Col xs={2}>
+            <Field
+              component={TextField}
+              label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.field.amount`} />}
+              name={getSubfieldName(41, 1, index)}
+            />
+          </Col>
+          <Col xs={2}>
+            <Field
+              component={TypeToggle}
+              label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.field.type`} />}
+              name={getSubfieldName(41, 2, index)}
+            />
+          </Col>
+          <Col xs={3}>
+            <AcceptedValuesField
+              component={TextField}
+              name={getSubfieldName(41, 3, index)}
+              label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.field.relationToTotal`} />}
+              optionValue="value"
+              optionLabel="label"
+              isRemoveValueAllowed
+              wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
+              acceptedValuesList={relationToTotalList}
+              okapi={okapi}
+            />
+          </Col>
+          <Col xs={2}>
+            <Field
+              component={Checkbox}
+              vertical
+              label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.field.exportToAccounting`} />}
+              name={getSubfieldName(41, 4, index)}
+            />
+          </Col>
+        </Row>
+      </Card>
+    );
+  };
+
+  return (
+    <Accordion
+      id="invoice-line-adjustments"
+      label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceLineAdjustments.section`} />}
+    >
+      <Row left="xs">
+        <Col
+          xs={12}
+        >
+          <RepeatableField
+            fields={lineAdjustments}
+            addLabel={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.invoiceAdjustments.adjustments.addLabel`} />}
+            onAdd={() => onAdd(lineAdjustments, 'lineAdjustments', 41, initialFields, setReferenceTables, 'order')}
+            onRemove={null}
+            renderField={renderLineAdjustment}
+          />
+        </Col>
+      </Row>
+    </Accordion>
+  );
+};
+
+InvoiceLineAdjustments.propTypes = {
+  lineAdjustments: PropTypes.arrayOf(mappingProfileSubfieldShape).isRequired,
+  initialFields: PropTypes.object.isRequired,
+  setReferenceTables: PropTypes.func.isRequired,
+  okapi: okapiShape.isRequired,
+};