+++ src/utils/createOptionsList.js
@@ -1,4 +1,4 @@
-export const createOptionsList = (array, formatMessage) => array.map(option => ({
+export const createOptionsList = (array, formatMessage, customLabel) => array.map(option => ({
   value: option.value,
-  label: formatMessage({ id: option.label }),
+  label: formatMessage({ id: option[customLabel] || option.label }),
 }));